+++ doc/examples/animations/lib/main.dart
@@ -2,21 +2,22 @@ import 'package:flame/gestures.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
-import 'package:flame/animation.dart' as flame_animation;
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/vector2.dart';
+import 'package:flame/sprite_animation.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Flame.images.loadAll(['creature.png', 'chopper.png']);
 
-  final Size size = await Flame.util.initialDimensions();
+  final Vector2 size = await Flame.util.initialDimensions();
   final game = MyGame(size);
   runApp(game.widget);
 }
 
 class MyGame extends BaseGame with TapDetector {
-  final animation = flame_animation.Animation.sequenced(
+  final animation = SpriteAnimation.sequenced(
     'chopper.png',
     4,
     textureWidth: 48,
@@ -28,7 +29,8 @@ class MyGame extends BaseGame with TapDetector {
   void addAnimation(double x, double y) {
     const textureWidth = 291.0;
     const textureHeight = 178.0;
-    final animationComponent = AnimationComponent.sequenced(
+
+    final animationComponent = SpriteAnimationComponent.sequenced(
       291,
       178,
       'creature.png',
@@ -40,6 +42,7 @@ class MyGame extends BaseGame with TapDetector {
       loop: false,
       destroyOnFinish: true,
     );
+
     animationComponent.x = x - textureWidth / 2;
     animationComponent.y = y - textureHeight / 2;
 
@@ -51,17 +54,20 @@ class MyGame extends BaseGame with TapDetector {
     addAnimation(evt.globalPosition.dx, evt.globalPosition.dy);
   }
 
-  MyGame(Size screenSize) {
+  MyGame(Vector2 screenSize) {
     size = screenSize;
 
     const s = 100.0;
-    final animationComponent = AnimationComponent(s, s, animation);
-    animationComponent.x = size.width / 2 - s;
+    final animationComponent = SpriteAnimationComponent(s, s, animation);
+    animationComponent.x = size.x / 2 - s;
     animationComponent.y = s;
 
-    final reversedAnimationComponent =
-        AnimationComponent(s, s, animation.reversed());
-    reversedAnimationComponent.x = size.width / 2;
+    final reversedAnimationComponent = SpriteAnimationComponent(
+      s,
+      s,
+      animation.reversed(),
+    );
+    reversedAnimationComponent.x = size.x / 2;
     reversedAnimationComponent.y = s;
 
     add(animationComponent);