+++ doc/examples/aseprite/lib/main.dart
@@ -1,30 +1,31 @@
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
-import 'package:flame/animation.dart' as flame_animation;
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/sprite_animation.dart';
+import 'package:flame/vector2.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  final Size size = await Flame.util.initialDimensions();
+  final Vector2 size = await Flame.util.initialDimensions();
   runApp(MyGame(size).widget);
 }
 
 class MyGame extends BaseGame {
-  MyGame(Size screenSize) {
+  MyGame(Vector2 screenSize) {
     size = screenSize;
     _start();
   }
 
   void _start() async {
-    final animation = await flame_animation.Animation.fromAsepriteData(
+    final animation = await SpriteAnimation.fromAsepriteData(
       'chopper.png',
       'chopper.json',
     );
-    final animationComponent = AnimationComponent(200, 200, animation);
+    final animationComponent = SpriteAnimationComponent(200, 200, animation);
 
-    animationComponent.x = (size.width / 2) - 100;
-    animationComponent.y = (size.height / 2) - 100;
+    animationComponent.x = (size.x / 2) - 100;
+    animationComponent.y = (size.y / 2) - 100;
 
     add(animationComponent);
   }