+++ doc/examples/effects/combined_effects/lib/square.dart
@@ -1,5 +1,5 @@
 import 'package:flame/anchor.dart';
-import 'package:flame/components/component.dart';
+import 'package:flame/components/position_component.dart';
 
 import 'dart:ui';
 
@@ -17,8 +17,7 @@ class Square extends PositionComponent {
 
   @override
   void render(Canvas canvas) {
-    prepareCanvas(canvas);
-    final rect = Rect.fromLTWH(0, 0, width, height);
-    canvas.drawRect(rect, _paint);
+    super.render(canvas);
+    canvas.drawRect(toOriginRect(), _paint);
   }
 }