+++ lib/particles/animation_particle.dart
@@ -1,18 +1,18 @@
 import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
+import 'package:vector_math/vector_math_64.dart';
 
-import '../animation.dart';
+import '../sprite_animation.dart';
 import '../particle.dart';
-import '../position.dart';
 
-class AnimationParticle extends Particle {
-  final Animation animation;
-  final Position size;
+class SpriteAnimationParticle extends Particle {
+  final SpriteAnimation animation;
+  final Vector2 size;
   final Paint overridePaint;
   final bool alignAnimationTime;
 
-  AnimationParticle({
+  SpriteAnimationParticle({
     @required this.animation,
     this.size,
     this.overridePaint,
@@ -36,7 +36,7 @@ class AnimationParticle extends Particle {
   void render(Canvas canvas) {
     animation.getSprite().renderCentered(
           canvas,
-          Position.empty(),
+          Vector2.zero(),
           overridePaint: overridePaint,
           size: size,
         );