+++ lib/util.dart
@@ -1,17 +1,10 @@
 import 'dart:async';
 import 'dart:ui';
 
+import 'package:flame/vector2.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart' as widgets;
-
-import 'animation.dart';
-import 'game/base_game.dart';
-import 'game/embedded_game_widget.dart';
-import 'sprite.dart';
-import 'components/animation_component.dart';
-import 'position.dart';
 
 /// Some utilities that did not fit anywhere else.
 ///
@@ -104,20 +97,22 @@ class Util {
   ///
   /// A best practice would be to implement there resize hooks on your game and components and don't use this at all.
   /// Make sure your components are able to render and update themselves for any possible screen size.
-  Future<Size> initialDimensions() async {
+  Future<Vector2> initialDimensions() async {
     // https://github.com/flutter/flutter/issues/5259
     // "In release mode we start off at 0x0 but we don't in debug mode"
-    return await Future<Size>(() {
+    return await Future<Vector2>(() {
       if (window.physicalSize.isEmpty) {
-        final completer = Completer<Size>();
+        final completer = Completer<Vector2>();
         window.onMetricsChanged = () {
           if (!window.physicalSize.isEmpty && !completer.isCompleted) {
-            completer.complete(window.physicalSize / window.devicePixelRatio);
+            completer.complete(Vector2Operations.fromSize(
+                window.physicalSize / window.devicePixelRatio));
           }
         };
         return completer.future;
       }
-      return window.physicalSize / window.devicePixelRatio;
+      return Vector2Operations.fromSize(
+          window.physicalSize / window.devicePixelRatio);
     });
   }
 
@@ -154,48 +149,9 @@ class Util {
   ///
   /// Some render methods don't allow to pass a offset.
   /// This method translate the canvas, draw what you want, and then translate back.
-  void drawWhere(Canvas c, Position p, void Function(Canvas) fn) {
+  void drawWhere(Canvas c, Vector2 p, void Function(Canvas) fn) {
     c.translate(p.x, p.y);
     fn(c);
     c.translate(-p.x, -p.y);
   }
-
-  /// Returns a regular Flutter widget representing this animation, rendered with the specified size.
-  ///
-  /// This actually creates an [EmbeddedGameWidget] with a [SimpleGame] whose only content is an [AnimationComponent] created from the provided [animation].
-  /// You can use this implementation as base to easily create your own widgets based on more complex games.
-  /// This is intended to be used by non-game apps that want to add a sprite sheet animation.
-  ///
-  @Deprecated('Use SpriteAnimation instead')
-  widgets.Widget animationAsWidget(Position size, Animation animation) {
-    return EmbeddedGameWidget(
-      BaseGame()..add(AnimationComponent(size.x, size.y, animation)),
-      size: size,
-    );
-  }
-
-  /// Returns a regular Flutter widget representing this sprite, rendered with the specified size.
-  ///
-  /// This will create a [CustomPaint] widget using a [CustomPainter] for rendering the [Sprite]
-  /// Be aware that the Sprite must have been loaded, otherwise it can't be rendered
-  ///
-  @Deprecated('Use SpriteWidget instead')
-  widgets.CustomPaint spriteAsWidget(Size size, Sprite sprite) =>
-      widgets.CustomPaint(size: size, painter: _SpriteCustomPainter(sprite));
-}
-
-class _SpriteCustomPainter extends widgets.CustomPainter {
-  final Sprite _sprite;
-
-  _SpriteCustomPainter(this._sprite);
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    if (_sprite.loaded()) {
-      _sprite.render(canvas, width: size.width, height: size.height);
-    }
-  }
-
-  @override
-  bool shouldRepaint(widgets.CustomPainter old) => false;
 }