+++ test/components/has_game_ref_test.dart
@@ -1,10 +1,8 @@
-import 'dart:ui';
-
 import 'package:flame/components/mixins/has_game_ref.dart';
 import 'package:flame/game/base_game.dart';
 import 'package:test/test.dart';
 
-import 'package:flame/components/component.dart';
+import 'package:flame/components/position_component.dart';
 
 class MyGame extends BaseGame {
   bool calledFoo = false;
@@ -14,9 +12,6 @@ class MyGame extends BaseGame {
 }
 
 class MyComponent extends PositionComponent with HasGameRef<MyGame> {
-  @override
-  void render(Canvas c) {}
-
   void foo() {
     gameRef.foo();
   }